JarInfer CLI tool and in-memory jar/aar processing (#185)
Added CLI tool for invoking on specified jar / aar file.
