+++ gradle/dependencies.gradle
@@ -20,6 +20,7 @@ def versions = [
     errorProne             : "2.1.1",
     support                : "27.0.2",
     wala                   : "1.4.3",
+    commonscli             : "1.4",
 ]
 
 def apt = [
@@ -41,6 +42,7 @@ def build = [
     wala                    : ["com.ibm.wala:com.ibm.wala.util:${versions.wala}",
                                "com.ibm.wala:com.ibm.wala.shrike:${versions.wala}",
                                "com.ibm.wala:com.ibm.wala.core:${versions.wala}"],
+    commonscli              : "commons-cli:commons-cli:${versions.commonscli}",
 
     // android stuff
     buildToolsVersion: "27.0.2",
