+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,3 +1,4 @@
+#Fri Jun 29 14:35:16 PDT 2018
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
 zipStoreBase=GRADLE_USER_HOME
