--- jar-infer/build.gradle
@@ -1,59 +1,0 @@
-plugins {
-    id "net.ltgt.errorprone" version "0.0.13"
-    id "java"
-    id "application"
-}
-// we use this config to get the path of the JDK 9 javac jar, to
-// stick it in the bootclasspath when running tests
-configurations.maybeCreate("epJavac")
-
-sourceCompatibility = 1.8
-version = '0.1'
-jar {
-    manifest {
-        attributes 'Implementation-Title': 'WALA Start',
-                   'Implementation-Version': version
-    }
-}
-
-repositories {
-    mavenCentral()
-    // uncomment if you want to use wala.dalvik or wala.scandroid
-    //maven { url "https://raw.github.com/msridhar/maven-jars/mvn-repo/" }
-}
-
-dependencies {
-    compile deps.build.wala
-    compile deps.build.guava
-    compile deps.build.commonsIO
-    compileOnly deps.build.errorProneCheckApi
-    compile deps.build.checkerDataflow
-
-    errorprone deps.build.errorProneCore
-
-    testCompile deps.test.junit4
-    testCompile(deps.build.errorProneTestHelpers) {
-        exclude group: "junit", module: "junit"
-    }
-    epJavac deps.build.errorProneCheckApi
-}
-
-if (project.hasProperty('mainClass')) {
-    mainClassName = project.getProperty('mainClass')
-} else {
-    // use a default
-    mainClassName = "com.ibm.wala.examples.drivers.definitelyDerefedParams"
-}
-
-test {
-    maxHeapSize = "1024m"
-    jvmArgs "-Xbootclasspath/p:${configurations.epJavac.asPath}"
-}
-
-run {
-    if(project.hasProperty('appArgs')){
-        args Eval.me(appArgs)
-    }
-}
-
-apply from: rootProject.file("gradle/gradle-mvn-push.gradle")
