+++ settings.gradle
@@ -4,5 +4,7 @@ include ':sample'
 include ':sample-app'
 include ':test-java-lib'
 include ':test-java-lib-jarinfer'
+include ':test-android-lib-jarinfer'
 include ':compile-bench'
-include ':jar-infer'
+include ':jar-infer-lib'
+include ':jar-infer-cli'
