+++ test-android-lib-jarinfer/src/main/java/org/utilities/StringUtils.java
@@ -1,0 +1,10 @@
+package org.utilities;
+
+import android.support.annotation.Nullable;
+
+public class StringUtils {
+
+  public static boolean isEmptyOrNull(@Nullable final CharSequence value) {
+    return value == null || value.length() == 0;
+  }
+}
